feat(ds): stack data structure
